@@ -51,8 +51,12 @@
     "@testing-library/react": "^12.1.3",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.20.43",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
+    "ajv": "^8.6.3",
     "babel-jest": "^27.5.1",
     "concurrently": "^7.0.0",
+    "eslint": "^8.9.0",
     "jasmine": "^3.5.0",
     "jest": "^27.5.1",
     "jsdom": "^19.0.0",
@@ -63,8 +67,18 @@
     "react-dom": "^17.0.2",
     "react-test-renderer": "^17.0.2",
     "ts-jest": "^27.1.3",
-    "typescript": "~4.3.2",
+    "typescript": "~4.6-rc",
     "zx": "^5.1.0"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{js,json,yml,yaml,css,scss,ts,tsx,md}": [
+      "prettier --write"
+    ]
   },
   "dependencies": {
     "helloworld-html-kodyfire": "file:packages/helloworld-html-kodyfire"
