@@ -27,7 +27,7 @@
     "@angular-devkit/schematics": "^12.2.3",
     "ajv": "^8.6.3",
     "fs-extra": "^10.0.0",
-    "kodyfire-core": "^0.0.56",
-    "typescript": "~4.3.2"
+    "kodyfire-core": "^0.0.66",
+    "typescript": "~4.6-rc"
   }
 }
