@@ -23,8 +23,8 @@
     "@angular-devkit/schematics": "^12.2.3",
     "ajv": "^8.6.3",
     "fs-extra": "^10.0.0",
-    "kodyfire-core": "^0.0.56",
-    "typescript": "~4.3.2"
+    "kodyfire-core": "^0.0.66",
+    "typescript": "~4.6-rc"
   },
   "devDependencies": {
     "@types/node": "^17.0.14"
@@ -34,8 +34,5 @@
     "type": "web",
     "version": "5"
   },
-  "license": "ISC",
-  "publishConfig": {
-    "@noqta:registry": "https://gitlab.noqta.tn/api/v4/projects/163/packages/npm/"
-  }
+  "license": "ISC"
 }
