@@ -19,7 +19,7 @@
   "license": "ISC",
   "dependencies": {
     "@angular-devkit/schematics": "^12.2.3",
-    "typescript": "~4.3.2"
+    "typescript": "~4.6-rc"
   },
   "devDependencies": {
     "@types/node": "^17.0.14"
