@@ -28,7 +28,7 @@
     "fs-extra": "^10.0.0",
     "kodyfire-core": "^0.0.56",
     "moment": "^2.29.1",
-    "typescript": "~4.3.2"
+    "typescript": "~4.6-rc"
   },
   "devDependencies": {
     "@types/pluralize": "0.0.29"
