@@ -10,8 +10,8 @@ export class Generator implements IGenerator {
   }
   generate(content: any) {
     // for every concept in concepts list
-    for (let [key] of this.technology.concepts) {
-      for (let data of content[key]) {
+    for (const [key] of this.technology.concepts) {
+      for (const data of content[key]) {
         // do apropriate action
         this.output = this.technology.concepts.get(key)?.generate(data);
       }
