@@ -1,6 +1,3 @@
-declare const _default: (
-  uischema: import('@jsonforms/core').UISchemaElement,
-  schema: import('@jsonforms/core').JsonSchema
-) => number;
+declare const _default: (uischema: import("@jsonforms/core").UISchemaElement, schema: import("@jsonforms/core").JsonSchema) => number;
 export default _default;
 //# sourceMappingURL=ratingControlTester.d.ts.map
