@@ -1,8 +1,6 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
-const core_1 = require('@jsonforms/core');
-exports.default = core_1.rankWith(
-  3, //increase rank as needed
-  core_1.scopeEndsWith('rating')
-);
+
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@jsonforms/core");
+exports.default = core_1.rankWith(3, //increase rank as needed
+core_1.scopeEndsWith('rating'));
 //# sourceMappingURL=ratingControlTester.js.map
