@@ -1,95 +1,23 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
-exports.KodyWorkflow =
-  exports.underscore =
-  exports.decamelize =
-  exports.dasherize =
-  exports.camelize =
-  exports.capitalize =
-  exports.ActionList =
-  exports.Source =
-  exports.Technology =
-  exports.BaseKody =
-  exports.Package =
-  exports.Runner =
-    void 0;
-const lib_1 = require('./lib');
-Object.defineProperty(exports, 'Runner', {
-  enumerable: true,
-  get: function () {
-    return lib_1.Runner;
-  },
-});
-Object.defineProperty(exports, 'Package', {
-  enumerable: true,
-  get: function () {
-    return lib_1.Package;
-  },
-});
-const kody_1 = require('./typings/kody');
-Object.defineProperty(exports, 'BaseKody', {
-  enumerable: true,
-  get: function () {
-    return kody_1.BaseKody;
-  },
-});
-const actionList_1 = require('./typings/actionList');
-Object.defineProperty(exports, 'ActionList', {
-  enumerable: true,
-  get: function () {
-    return actionList_1.ActionList;
-  },
-});
-const technology_1 = require('./typings/technology');
-Object.defineProperty(exports, 'Technology', {
-  enumerable: true,
-  get: function () {
-    return technology_1.Technology;
-  },
-});
-const concept_1 = require('./typings/concept');
-Object.defineProperty(exports, 'Source', {
-  enumerable: true,
-  get: function () {
-    return concept_1.Source;
-  },
-});
-const workflow_1 = require('./typings/workflow');
-Object.defineProperty(exports, 'KodyWorkflow', {
-  enumerable: true,
-  get: function () {
-    return workflow_1.KodyWorkflow;
-  },
-});
-const strings_1 = require('@angular-devkit/core/src/utils/strings');
-Object.defineProperty(exports, 'capitalize', {
-  enumerable: true,
-  get: function () {
-    return strings_1.capitalize;
-  },
-});
-Object.defineProperty(exports, 'camelize', {
-  enumerable: true,
-  get: function () {
-    return strings_1.camelize;
-  },
-});
-Object.defineProperty(exports, 'dasherize', {
-  enumerable: true,
-  get: function () {
-    return strings_1.dasherize;
-  },
-});
-Object.defineProperty(exports, 'decamelize', {
-  enumerable: true,
-  get: function () {
-    return strings_1.decamelize;
-  },
-});
-Object.defineProperty(exports, 'underscore', {
-  enumerable: true,
-  get: function () {
-    return strings_1.underscore;
-  },
-});
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.KodyWorkflow = exports.underscore = exports.decamelize = exports.dasherize = exports.camelize = exports.capitalize = exports.ActionList = exports.Source = exports.Technology = exports.BaseKody = exports.Package = exports.Runner = void 0;
+const lib_1 = require("./lib");
+Object.defineProperty(exports, "Runner", { enumerable: true, get: function () { return lib_1.Runner; } });
+Object.defineProperty(exports, "Package", { enumerable: true, get: function () { return lib_1.Package; } });
+const kody_1 = require("./typings/kody");
+Object.defineProperty(exports, "BaseKody", { enumerable: true, get: function () { return kody_1.BaseKody; } });
+const actionList_1 = require("./typings/actionList");
+Object.defineProperty(exports, "ActionList", { enumerable: true, get: function () { return actionList_1.ActionList; } });
+const technology_1 = require("./typings/technology");
+Object.defineProperty(exports, "Technology", { enumerable: true, get: function () { return technology_1.Technology; } });
+const concept_1 = require("./typings/concept");
+Object.defineProperty(exports, "Source", { enumerable: true, get: function () { return concept_1.Source; } });
+const workflow_1 = require("./typings/workflow");
+Object.defineProperty(exports, "KodyWorkflow", { enumerable: true, get: function () { return workflow_1.KodyWorkflow; } });
+const strings_1 = require("@angular-devkit/core/src/utils/strings");
+Object.defineProperty(exports, "capitalize", { enumerable: true, get: function () { return strings_1.capitalize; } });
+Object.defineProperty(exports, "camelize", { enumerable: true, get: function () { return strings_1.camelize; } });
+Object.defineProperty(exports, "dasherize", { enumerable: true, get: function () { return strings_1.dasherize; } });
+Object.defineProperty(exports, "decamelize", { enumerable: true, get: function () { return strings_1.decamelize; } });
+Object.defineProperty(exports, "underscore", { enumerable: true, get: function () { return strings_1.underscore; } });
 //# sourceMappingURL=index.js.map
