@@ -1,18 +1,8 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.Runner = exports.Package = void 0;
-const package_1 = require('./package');
-Object.defineProperty(exports, 'Package', {
-  enumerable: true,
-  get: function () {
-    return package_1.Package;
-  },
-});
-const runner_1 = require('./runner');
-Object.defineProperty(exports, 'Runner', {
-  enumerable: true,
-  get: function () {
-    return runner_1.Runner;
-  },
-});
+const package_1 = require("./package");
+Object.defineProperty(exports, "Package", { enumerable: true, get: function () { return package_1.Package; } });
+const runner_1 = require("./runner");
+Object.defineProperty(exports, "Runner", { enumerable: true, get: function () { return runner_1.Runner; } });
 //# sourceMappingURL=index.js.map
