@@ -1,3 +1,3 @@
 'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+Object.defineProperty(exports, "__esModule", { value: true });
 //# sourceMappingURL=template.js.map
