@@ -1,3 +1,4 @@
 import { ActionList as BaseActionList } from '@angular-devkit/schematics';
-export declare class ActionList extends BaseActionList {}
+export declare class ActionList extends BaseActionList {
+}
 //# sourceMappingURL=actionList.d.ts.map
