@@ -1,7 +1,8 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.ActionList = void 0;
-const schematics_1 = require('@angular-devkit/schematics');
-class ActionList extends schematics_1.ActionList {}
+const schematics_1 = require("@angular-devkit/schematics");
+class ActionList extends schematics_1.ActionList {
+}
 exports.ActionList = ActionList;
 //# sourceMappingURL=actionList.js.map
