@@ -1,5 +1,5 @@
 export interface IArgument {
-  name: string;
-  value: string;
+    name: string;
+    value: string;
 }
 //# sourceMappingURL=argument.d.ts.map
