@@ -1,8 +1,8 @@
 export declare enum ArtifactType {
-  file = 0,
+    file = 0
 }
 export interface IArtifact {
-  name: string;
-  type: ArtifactType;
+    name: string;
+    type: ArtifactType;
 }
 //# sourceMappingURL=artifact.d.ts.map
