@@ -1,8 +1,8 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.ArtifactType = void 0;
 var ArtifactType;
 (function (ArtifactType) {
-  ArtifactType[(ArtifactType['file'] = 0)] = 'file';
-})((ArtifactType = exports.ArtifactType || (exports.ArtifactType = {})));
+    ArtifactType[ArtifactType["file"] = 0] = "file";
+})(ArtifactType = exports.ArtifactType || (exports.ArtifactType = {}));
 //# sourceMappingURL=artifact.js.map
