@@ -1,8 +1,8 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.Source = void 0;
 var Source;
 (function (Source) {
-  Source[(Source['Template'] = 0)] = 'Template';
-})((Source = exports.Source || (exports.Source = {})));
+    Source[Source["Template"] = 0] = "Template";
+})(Source = exports.Source || (exports.Source = {}));
 //# sourceMappingURL=concept.js.map
