@@ -1,28 +1,28 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.BaseKody = void 0;
 class BaseKody {
-  get data() {
-    throw new Error('Method not implemented.');
-  }
-  get errors() {
-    throw new Error('Method not implemented.');
-  }
-  parse(_content) {
-    throw new Error('Method not implemented.');
-  }
-  read(_source) {
-    throw new Error('Method not implemented.');
-  }
-  generate(_content) {
-    throw new Error('Method not implemented.');
-  }
-  whereami() {
-    throw new Error('Method not implemented.');
-  }
-  whoami() {
-    throw new Error('Method not implemented.');
-  }
+    get data() {
+        throw new Error('Method not implemented.');
+    }
+    get errors() {
+        throw new Error('Method not implemented.');
+    }
+    parse(_content) {
+        throw new Error('Method not implemented.');
+    }
+    read(_source) {
+        throw new Error('Method not implemented.');
+    }
+    generate(_content) {
+        throw new Error('Method not implemented.');
+    }
+    whereami() {
+        throw new Error('Method not implemented.');
+    }
+    whoami() {
+        throw new Error('Method not implemented.');
+    }
 }
 exports.BaseKody = BaseKody;
 //# sourceMappingURL=kody.js.map
