@@ -1,7 +1,7 @@
 export interface IPackage {
-  id: string;
-  name?: string;
-  type: string;
-  version: string;
+    id: string;
+    name?: string;
+    type: string;
+    version: string;
 }
 //# sourceMappingURL=package.d.ts.map
