@@ -4,12 +4,12 @@ import { IValidator } from './validator';
  * @alpha
  */
 export interface IParser {
-  data: any;
-  requiresExtract: boolean;
-  requiresTansform: boolean;
-  requiresLoad: boolean;
-  validator: IValidator;
-  reader(filename: string): any;
-  parse(data: any): any;
+    data: any;
+    requiresExtract: boolean;
+    requiresTansform: boolean;
+    requiresLoad: boolean;
+    validator: IValidator;
+    reader(filename: string): any;
+    parse(data: any): any;
 }
 //# sourceMappingURL=parser.d.ts.map
