@@ -1,12 +1,12 @@
-'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 exports.Technology = void 0;
-const actionList_1 = require('./actionList');
+const actionList_1 = require("./actionList");
 class Technology {
-  constructor() {
-    this.actions = new actionList_1.ActionList();
-    this.concepts = new Map();
-  }
+    constructor() {
+        this.actions = new actionList_1.ActionList();
+        this.concepts = new Map();
+    }
 }
 exports.Technology = Technology;
 //# sourceMappingURL=technology.js.map
