@@ -1,8 +1,8 @@
 export interface Template {
-  id: any;
-  name: any;
-  type: any;
-  version: any;
-  enabled: any;
+    id: any;
+    name: any;
+    type: any;
+    version: any;
+    enabled: any;
 }
 //# sourceMappingURL=template.d.ts.map
