@@ -1,6 +1,6 @@
 export interface IValidator {
-  rules: any;
-  errors: any;
-  validate(data: any): boolean;
+    rules: any;
+    errors: any;
+    validate(data: any): boolean;
 }
 //# sourceMappingURL=validator.d.ts.map
